patch: navigation
update all navigation to App, SignInScreen and SignUpScreen
